Update main.c
description of compiler warning:

gcc  -std=c99 -ggdb3 -Wall -Werror -Wextra -pedantic  -c main.c -o main.o
main.c: In function ‘main’:
main.c:305:26: error: comparison between signed and unsigned integer expressions [-Werror=sign-compare]
     for (size_t i = 0; i < pass_cnt; ++i) {
                          ^
cc1: all warnings being treated as errors
Makefile:23: recipe for target 'main.o' failed
make: *** [main.o] Error 1

+++ main.c
@@ -302,7 +302,7 @@ main(int argc, char **argv)
         exit(EXIT_FAILURE);
     }
 
-    for (size_t i = 0; i < pass_cnt; ++i) {
+    for (long i = 0; i < pass_cnt; ++i) {
         char *pass;
 
         /* We needn't add pass to tree since we free ASAP */
